requirements.txt
numpy
scikit-learn
mysql-connector-python